requirements.txt
aioredis==2.0.1
async-timeout==4.0.2
bidict==0.22.0
cffi==1.15.1
cleo==1.0.0a4
click==8.1.3
colorama==0.4.5
crashtest==0.3.1
Deprecated==1.2.13
dnspython==2.2.1
eventlet==0.33.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-SocketIO==5.2.0
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
hiredis==2.0.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
packaging==21.3
pandas==1.4.3
pptree==3.1
pycparser==2.21
pydantic==1.9.2
pylev==1.4.0
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.4
python-socketio==5.7.1
python-ulid==1.1.0
pytz==2022.2.1
redis==4.3.4
redis-om==0.0.27
six==1.16.0
types-redis==4.3.14
types-six==1.16.18
typing-extensions==4.3.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
